Correção para compilação em Lazarus 64 bits
git-svn-id: svn://svn.code.sf.net/p/acbr/code/trunk2@20801 6e92efe7-b92a-0410-a9ec-f9e4e41bb3a6

+++ Fontes/ACBrComum/ACBrImage.pas
@@ -219,7 +219,7 @@ begin
   if RealWidth < WidthExtended then
     bSizePixelArr := (BytesPerWidth * bHeight) ;
 
-  StreamLastPos := min(bPixelOffset + bSizePixelArr - 1, ABMPStream.Size-1);
+  StreamLastPos := min(Int64(bPixelOffset + bSizePixelArr - 1), ABMPStream.Size-1);
 
   while (StreamLastPos >= bPixelOffset) do
   begin
